Move fan control functions to separate .h/.cpp files
