+++ speeduino/board_stm32_official.cpp
@@ -8,6 +8,7 @@
 #include "comms_secondary.h"
 #include "scheduler_ignition_controller.h"
 #include "scheduler_fuel_controller.h"
+#include "src/controllers/fan/fanController.h"
 
 #if defined(BOARD_FCR_MICRO_F4)
 extern "C" void __real_pinMode(uint8_t pin, uint8_t mode);
