+++ speeduino/speeduino.ino
@@ -58,6 +58,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include "src/controllers/fuelPump/fuelPumpController.h"
 #include "scheduler_fuel_controller.h"
 #include "src/controllers/tsCommand/tsCommandController.h"
+#include "src/controllers/fan/fanController.h"
 
 #define CRANK_RUN_HYSTER    15
