+++ speeduino/src/controllers/fan/fanController.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <stdint.h>
+
+void initialiseFan(uint8_t fanPin);
+
+void fanControl(void);
+
+void fanOn(void);
+void fanOff(void);
+
+void fanInterrupt(void);
